+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [Unreleased]
 
+- Allow optional workingDir for ctx.RunCommand
+
 ## [1.0.3] - 2023-11-08
 
-- Fix command catupure new line issue
+- Fix command capture new line issue
 
 ## [1.0.2] - 2023-10-19